+++ asyncgit/src/sync/hooks.rs
@@ -74,14 +74,40 @@ pub fn hooks_prepare_commit_msg(
 
 #[cfg(test)]
 mod tests {
+	use std::{ffi::OsString, io::Write as _, path::Path};
+
+	use git2::Repository;
+	use tempfile::TempDir;
+
 	use super::*;
-	use crate::sync::tests::repo_init;
-	use std::fs::File;
-	use std::io::Write;
-	use std::path::Path;
+	use crate::sync::tests::repo_init_with_prefix;
+
+	fn repo_init() -> Result<(TempDir, Repository)> {
+		let mut os_string: OsString = OsString::new();
+
+		os_string.push("gitui $# ' ");
+
+		#[cfg(target_os = "linux")]
+		{
+			use std::os::unix::ffi::OsStrExt;
+
+			const INVALID_UTF8: &[u8] = b"\xED\xA0\x80";
+
+			os_string.push(std::ffi::OsStr::from_bytes(INVALID_UTF8));
+
+			assert!(os_string.to_str().is_none());
+		}
+
+		os_string.push(" ");
+
+		repo_init_with_prefix(os_string)
+	}
 
 	fn create_hook_in_path(path: &Path, hook_script: &[u8]) {
-		File::create(path).unwrap().write_all(hook_script).unwrap();
+		std::fs::File::create(path)
+			.unwrap()
+			.write_all(hook_script)
+			.unwrap();
 
 		#[cfg(unix)]
 		{
@@ -102,7 +128,7 @@ mod tests {
 		let hook = b"#!/bin/sh
 	echo 'rejected'
 	exit 1
-	        ";
+			";
 
 		git2_hooks::create_hook(
 			&repo,
@@ -113,9 +139,7 @@ mod tests {
 		let subfolder = root.join("foo/");
 		std::fs::create_dir_all(&subfolder).unwrap();
 
-		let res =
-			hooks_post_commit(&subfolder.to_str().unwrap().into())
-				.unwrap();
+		let res = hooks_post_commit(&subfolder.into()).unwrap();
 
 		assert_eq!(
 			res,
@@ -131,16 +155,12 @@ mod tests {
 	fn test_pre_commit_workdir() {
 		let (_td, repo) = repo_init().unwrap();
 		let root = repo.workdir().unwrap();
-		let repo_path: &RepoPath =
-			&root.as_os_str().to_str().unwrap().into();
-		let workdir =
-			crate::sync::utils::repo_work_dir(repo_path).unwrap();
+		let repo_path: &RepoPath = &root.to_path_buf().into();
 
 		let hook = b"#!/bin/sh
-	echo $(pwd)
+	echo \"$(pwd)\"
 	exit 1
-	        ";
-
+		";
 		git2_hooks::create_hook(
 			&repo,
 			git2_hooks::HOOK_PRE_COMMIT,
@@ -149,8 +169,9 @@ mod tests {
 		let res = hooks_pre_commit(repo_path).unwrap();
 		if let HookResult::NotOk(res) = res {
 			assert_eq!(
-				std::path::Path::new(res.trim_end()),
-				std::path::Path::new(&workdir)
+				res.trim_end().trim_end_matches('/'),
+				// TODO: fix if output isn't utf8.
+				root.to_string_lossy().trim_end_matches('/'),
 			);
 		} else {
 			assert!(false);
@@ -163,10 +184,10 @@ mod tests {
 		let root = repo.workdir().unwrap();
 
 		let hook = b"#!/bin/sh
-	echo 'msg' > $1
+	echo 'msg' > \"$1\"
 	echo 'rejected'
 	exit 1
-	        ";
+		";
 
 		git2_hooks::create_hook(
 			&repo,
@@ -178,11 +199,8 @@ mod tests {
 		std::fs::create_dir_all(&subfolder).unwrap();
 
 		let mut msg = String::from("test");
-		let res = hooks_commit_msg(
-			&subfolder.to_str().unwrap().into(),
-			&mut msg,
-		)
-		.unwrap();
+		let res =
+			hooks_commit_msg(&subfolder.into(), &mut msg).unwrap();
 
 		assert_eq!(
 			res,
@@ -199,7 +217,7 @@ mod tests {
 		let root = repo.workdir().unwrap();
 		let mut config = repo.config().unwrap();
 
-		const HOOKS_DIR: &'static str = "my_hooks";
+		const HOOKS_DIR: &str = "my_hooks";
 		config.set_str("core.hooksPath", HOOKS_DIR).unwrap();
 
 		let hook = b"#!/bin/sh
@@ -213,7 +231,7 @@ mod tests {
 
 		let mut msg = String::from("test");
 		let res = hooks_commit_msg(
-			&hooks_folder.to_str().unwrap().into(),
+			&hooks_folder.to_path_buf().into(),
 			&mut msg,
 		)
 		.unwrap();